cr feedback for provenance runBuilder testing - rename TypeAheadCombo, add retry/synchronization to file upload input
